# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Library",
    "description": "Contain all api from kulony backend"
  },
  "servers": [
    {
      "url": "https://kulony-backend.herokuapp.com/"
    }
  ],
  "paths": {
...
    "/api/user/edit_profile": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "แก้ไข Profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/{user_id}/profile": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "ขอข้อมูล User profile",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/following_topic": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "ขอ topic ที่ user follow ไว้",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/follow_topic/{topic_id}": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "ส่ง Topic ที่ user ต้องการ Follow",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/unfollow_topic/{topic_id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Unfollow topic ที่ user follow ไว้",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/user_like_post": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "ขอ post ที่ user like ไว้",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/mypost": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "ขอ Post ที่ user เป็นคนเขียน",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/topic/{catagory}": {
      "get": {
        "tags": [
          "Topic/Catagory"
        ],
        "description": "ขอ Topic ของ Catagory นั้น",
        "parameters": [
          {
            "name": "catagory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/topic/get_all_catagory_topic": {
      "get": {
        "tags": [
          "Topic/Catagory"
        ],
        "description": "ขอ Topic และ Catagory ทั้งหมด",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/topic/create_topic": {
      "post": {
        "tags": [
          "Topic/Catagory"
        ],
        "description": "สร้าง Topic ใหม่",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/topic/request_topic": {
      "post": {
        "tags": [
          "Topic/Catagory"
        ],
        "description": "ส่ง Request topic",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}